requirements.txt
yt-dlp
moviepy
openai-whisper
langchain-ollama